# finbot/apps/appwsrv/appwsrv.openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Finbot application service",
        "description": "API documentation for appwsrv",
        "version": "0.1.0"
    },
    "paths": {
...
        "/accounts/": {
            "post": {
                "tags": [
                    "User accounts"
                ],
                "summary": "Create User Account",
                "operationId": "create_user_account",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserAccountRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateUserAccountResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/accounts/{user_account_id}/": {
            "get": {
                "tags": [
                    "User accounts"
                ],
                "summary": "Get User Account",
                "operationId": "get_user_account",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "user_account_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "User Account Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetUserAccountResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/accounts/{user_account_id}/settings/": {
            "get": {
                "tags": [
                    "User accounts"
                ],
                "summary": "Get User Account Settings",
                "operationId": "get_user_account_settings",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "user_account_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "User Account Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetUserAccountSettingsResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/accounts/{user_account_id}/linked_accounts/": {
            "get": {
                "tags": [
                    "Linked accounts"
                ],
                "summary": "Get Linked Accounts",
                "description": "Get linked accounts",
                "operationId": "get_user_account_linked_accounts",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "user_account_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "User Account Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetLinkedAccountsResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Linked accounts"
                ],
                "summary": "Link New Account",
                "description": "Link new account",
                "operationId": "link_new_account",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "user_account_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "User Account Id"
                        }
                    },
                    {
                        "name": "validate",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": true,
                            "title": "Validate"
                        }
                    },
                    {
                        "name": "persist",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": true,
                            "title": "Persist"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LinkAccountRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LinkAccountResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/accounts/{user_account_id}/linked_accounts/{linked_account_id}/": {
            "get": {
                "tags": [
                    "Linked accounts"
                ],
                "summary": "Get Linked Account",
                "description": "Get linked account",
                "operationId": "get_linked_account",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "user_account_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "User Account Id"
                        }
                    },
                    {
                        "name": "linked_account_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Linked Account Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetLinkedAccountResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Linked accounts"
                ],
                "summary": "Delete Linked Account",
                "description": "Delete linked account",
                "operationId": "delete_linked_account",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "user_account_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "User Account Id"
                        }
                    },
                    {
                        "name": "linked_account_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Linked Account Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteLinkedAccountResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/accounts/{user_account_id}/linked_accounts/{linked_account_id}/metadata/": {
            "put": {
                "tags": [
                    "Linked accounts"
                ],
                "summary": "Update Linked Account Metadata",
                "description": "Update linked account metadata",
                "operationId": "update_linked_account_metadata",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "user_account_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "User Account Id"
                        }
                    },
                    {
                        "name": "linked_account_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Linked Account Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateLinkedAccountMetadataRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UpdateLinkedAccountMetadataResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/accounts/{user_account_id}/linked_accounts/{linked_account_id}/credentials/": {
            "put": {
                "tags": [
                    "Linked accounts"
                ],
                "summary": "Update Linked Account Credentials",
                "description": "Update linked account credentials",
                "operationId": "update_linked_account_credentials",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "user_account_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "User Account Id"
                        }
                    },
                    {
                        "name": "linked_account_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Linked Account Id"
                        }
                    },
                    {
                        "name": "validate",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": true,
                            "title": "Validate"
                        }
                    },
                    {
                        "name": "persist",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": true,
                            "title": "Persist"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateLinkedAccountCredentialsRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UpdateLinkedAccountCredentialsResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/formatting_rules/accounts/": {
            "get": {
                "tags": [
                    "Formatting rules"
                ],
                "summary": "Get Accounts Formatting Rules",
                "description": "Get accounts formatting rules",
                "operationId": "get_accounts_formatting_rules",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetAccountsFormattingRulesResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "CreateUserAccountRequest": {
                "properties": {
                    "email": {
                        "type": "string",
                        "title": "Email"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "title": "Password",
                        "writeOnly": true
                    },
                    "full_name": {
                        "type": "string",
                        "title": "Full Name"
                    },
                    "settings": {
                        "$ref": "#/components/schemas/UserAccountCreationSettings"
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                    "email",
                    "password",
                    "full_name",
                    "settings"
                ],
                "title": "CreateUserAccountRequest"
            },
            "CreateUserAccountResponse": {
                "properties": {
                    "user_account": {
                        "$ref": "#/components/schemas/UserAccount"
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                    "user_account"
                ],
                "title": "CreateUserAccountResponse"
            },
...
            "GetLinkedAccountsResponse": {
                "properties": {
                    "linked_accounts": {
                        "items": {
                            "$ref": "#/components/schemas/LinkedAccount"
                        },
                        "type": "array",
                        "title": "Linked Accounts"
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                    "linked_accounts"
                ],
                "title": "GetLinkedAccountsResponse"
            },
...
            "GetUserAccountResponse": {
                "properties": {
                    "user_account": {
                        "$ref": "#/components/schemas/UserAccount"
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                    "user_account"
                ],
                "title": "GetUserAccountResponse"
            },
            "GetUserAccountSettingsResponse": {
                "properties": {
                    "settings": {
                        "$ref": "#/components/schemas/UserAccountSettings"
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                    "settings"
                ],
                "title": "GetUserAccountSettingsResponse"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "LinkAccountResponse": {
                "properties": {},
                "additionalProperties": false,
                "type": "object",
                "title": "LinkAccountResponse"
            },
...
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}